+++ example/lib/pencil_field_with_tools.dart
@@ -1,5 +1,4 @@
 import 'dart:developer';
-
 import 'package:flutter/material.dart';
 import 'package:line_awesome_flutter/line_awesome_flutter.dart';
 import 'package:pencil_field/pencil_field.dart';